requirements.txt
Django~=3.0.7
django-allauth~=0.42.0
Pillow~=7.1.2
django-crispy-forms==1.9.1
django-invitations==1.9.3
